Stop MERGE-PATHNAMES from defaulting to :NEWEST VERSION
Previously, MERGE-PATHNAMES would add a :NEWEST VERSION if none was
specified.  Although "mostly harmless", this behavior could sometimes
end up with pathname objects that were not EQUAL when referring to the
same file.  We change the behavior to leave the version component
unchanged unless the MERGE-PATHNAME routine is called with an explicit
version.

From the CLHS entry for MERGE-PATHNAMES
<http://www.lispworks.com/documentation/HyperSpec/Body/f_merge_.htm>:

    Constructs a pathname from pathname by filling in any unsupplied
    components with the corresponding values from default-pathname and
    default-version.

    Defaulting of pathname components is done by filling in components
    taken from another pathname. This is especially useful for cases
    such as a program that has an input file and an output
    file. Unspecified components of the output pathname will come from
    the input pathname […]

    If no version is supplied, default-version is used. If
    default-version is nil, the version component will remain
    unchanged.

There is currently no known use of the implementation on a filesystem
where VERSION has a valid interpretation.